remote_server/requirements.txt
grpcio
grpcio-tools
opencv-python-headless
mediapipe
numpy
